% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{contract_coaches}
\alias{contract_coaches}
\title{Data contract for cleaned coaches}
\usage{
contract_coaches(coaches, stop_on_fail = TRUE)
}
\arguments{
\item{coaches}{Cleaned coaches (see \code{\link[=clean_coaches]{clean_coaches()}}).}

\item{stop_on_fail}{Abort on failure (default \code{TRUE})?}
}
\value{
The interrogated pointblank agent, invisibly.
}
\description{
Data contract for cleaned coaches
}
