% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{contract_player_stats}
\alias{contract_player_stats}
\title{Data contract for cleaned long player stats}
\usage{
contract_player_stats(player_stats, stop_on_fail = TRUE)
}
\arguments{
\item{player_stats}{Cleaned long stats (see \code{\link[=clean_player_stats]{clean_player_stats()}}).}

\item{stop_on_fail}{Abort on failure (default \code{TRUE})?}
}
\value{
The interrogated pointblank agent, invisibly.
}
\description{
Data contract for cleaned long player stats
}
