% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contracts.R
\name{contract_roster}
\alias{contract_roster}
\title{Data contract for cleaned rosters}
\usage{
contract_roster(roster, stop_on_fail = TRUE)
}
\arguments{
\item{roster}{Cleaned roster (see \code{\link[=clean_roster]{clean_roster()}}).}

\item{stop_on_fail}{Abort on failure (default \code{TRUE})?}
}
\value{
The interrogated pointblank agent, invisibly.
}
\description{
Data contract for cleaned rosters
}
